+++ src/gui/glaxnimate_app.hpp
@@ -3,7 +3,6 @@
 #include "app/log/listener_stderr.hpp"
 #include "app/log/listener_store.hpp"
 
-
 namespace app::settings { class ShortcutSettings; }
 
 class GlaxnimateApp : public app::Application