+++ src/gui/widgets/shape_style/fill_style_widget.hpp
@@ -25,6 +25,8 @@ private slots:
 
     void property_changed(const model::BaseProperty* prop);
 
+    void clear_target_color();
+
 private:
     model::Fill* target = nullptr;
     utils::PseudoMutex updating;